Sadly SQL Server seems to perform better so far

+++ Database/tables.sql
@@ -1,38 +1,41 @@
-drop table company cascade;
-drop table fact;
-drop table price;
+if object_id('price') is not null
+	drop table price
+
+if object_id('fact') is not null
+	drop table fact
+
+if object_id('company') is not null
+	drop table company
 
 create table company
 (
 	cik int primary key,
 	symbol text not null,
 	name text not null
-);
+)
 
 create table fact
 (
-	id serial primary key,
-	cik int references company (cik) not null,
+	cik int references company (cik) on delete cascade not null,
 	name text not null,
 	end_date date not null,
-	value numeric not null,
+	value decimal(28) not null,
 	fiscal_year int,
 	fiscal_period char(2),
-	form text not null,
+	form varchar(7) not null,
 	filed date not null,
 	frame text
-);
+)
 
-create index fact_cik_filed_form_index on fact (cik, filed, form);
+create clustered index fact_cik_filed_form_index on fact (cik, filed, form)
 
 create table price
 (
-	id serial primary key,
-	cik int references company (cik) not null,
+	cik int references company (cik) on delete cascade not null,
 	date date not null,
-	open money not null,
+	open_price money not null,
 	high money not null,
 	low money not null,
-	close money not null,
+	close_price money not null,
 	volume bigint not null
-);
+)
